requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
python-dotenv==1.0.0
requests==2.31.0
opencv-python-headless==4.8.0.76
numpy==1.24.3
email-validator==2.0.0
Werkzeug==2.3.7
supabase==2.0.2
fer==22.4.0
gunicorn==21.2.0
